[debug-info] Add a new form of SDDbgOperand that guarantees SelectionDAG will emit a dbg value as an entry value.
Today, we rely solely on Live Debug Values to create entry values for
debuggers. In certain cases though SelectionDAG knows that a value must be an
entry value (e.x.: the swift async context argument where variables in async
functions are stored). In such cases, we have found that relying on Live Debug
Values is not reliable since Live Debug Values runs /after/ register
allocation. Live Debug Values understands that there is a specific known async
argument register (r14 on Intel), so if we spill for whatever reason, the
pattern matching will fail and *BAM* no entry value. This patch is the beginning
of a series of patches that fixes this issue by teaching SelectionDAG how to
always mark debug_values that are from the swift context as being entry values.

That being said, this patch begins that process by mechanically adding a new
kind of SDDbgOperand::ENTRYVALUEREG and implementes the parts of SelectionDAG
necessary for SelectionDAGBuilder to be able to use this new SDDbgOperand. This
just meant adding the type and teaching InstrEmitter how to handle it. In
InstrEmitter, I just followed the model of SDDbgOperand::VREG except that I also
add an entry value to the debug value expression.

In a subsequent patch, I am going to add in the SelectionDAGBuilder support. I
am just splitting this to make it easier to reason about/review.